fix promotion flash-sale

+++ delivery-serivce/src/main/java/com/okits02/delivery_serivce/dto/request/OrderRequest.java
@@ -13,6 +13,6 @@ public class OrderRequest {
     String orderId;
     String addressId;
     String userId;
-    BigDecimal totalCost;
+    Integer totalCost;
     List<OrderItemRequest> items;
 }
